+++ src/config/mod.rs
@@ -54,6 +54,7 @@ pub use crate::address::{
 pub const DEFAULT_CONFIG_FILE_NAME: &str = "picodata.yaml";
 pub(crate) const DEFAULT_SQL_PREEMPTION: bool = false;
 pub(crate) const DEFAULT_SQL_PREEMPTION_INTERVAL_US: u64 = 500;
+pub(crate) const DEFAULT_SQL_LOG: bool = false;
 
 pub use ::sql::ir::types::DomainType as SbroadType;
 
@@ -1954,6 +1955,11 @@ pub struct AlterSystemParameters {
     #[introspection(sbroad_type = SbroadType::String)]
     #[introspection(config_default = generate_secure_token())]
     pub jwt_secret: String,
+
+    /// Enables logging of SQL statements.
+    #[introspection(sbroad_type = SbroadType::Boolean)]
+    #[introspection(config_default = DEFAULT_SQL_LOG)]
+    pub sql_log: bool,
 }
 
 fn generate_secure_token() -> String {
@@ -2091,6 +2097,7 @@ pub struct DynamicConfigProviders {
     pub sql_motion_row_max: AtomicObserverProvider<i64>,
     pub sql_preemption: AtomicObserverProvider<bool>,
     pub sql_preemption_interval_us: AtomicObserverProvider<u64>,
+    pub sql_log: AtomicObserverProvider<bool>,
 }
 
 impl DynamicConfigProviders {
@@ -2102,6 +2109,7 @@ impl DynamicConfigProviders {
             sql_motion_row_max: AtomicObserverProvider::new(),
             sql_preemption: AtomicObserverProvider::new(),
             sql_preemption_interval_us: AtomicObserverProvider::new(),
+            sql_log: AtomicObserverProvider::new(),
         }
     }
 
@@ -2455,6 +2463,10 @@ pub fn apply_parameter(
             }
             res
         })?;
+    } else if name == system_parameter_name!(sql_log) {
+        let value = v.as_bool().expect("type is already checked");
+        // Cache the value.
+        DYNAMIC_CONFIG.sql_log.update(value);
     }
 
     Ok(())